[DEV]SEOS-1077 SizeOfRecord changed to uint64_t as size_t is different for arm32 and x86-64 platforms

+++ include/OS_ConfigServiceBackend.h
@@ -97,7 +97,7 @@ OS_ConfigServiceBackend_createFileBackend(
     OS_ConfigServiceBackend_FileName_t name,
     hPartition_t phandle,
     unsigned int numberOfRecords,
-    size_t sizeOfRecord);
+    uint64_t sizeOfRecord);
 
 // Initialize the given backend object with the backend layout retrieved from
 // the given file.
